Assorted Fixes.
Fixed off-center background chain object in "Stairway to Hell".
Changed Stairway to Hell crate sizes to be uniform.
Added background objects in Stairway to Hell and Maze Runner to an empty object to make organization easier.
Fixed a visual glitch in Laser Lane that caused the laser beams to appear in front of the projectors.